Tweak the sequence API
Make the sequence API a bit more lenient, and tell it to ignore invalid
sequence identifiers if more than one identifier is being asked for,
i.e. in the multi sequence api, which the bulk feature download uses.

If a single invalid identifier is asked for, a 404 is thrown. Otherwise,
the subset of information for valid identifiers is returned.