[BI-1189] - Change seasons POST endpoint to comply with BrAPI spec
Original implementation was a proxy to the GET endpoint, and just returned the seasons that existed.  Breedbase doesn't save seasons separate from a study, so the new implementation creates a mock season with the seasonDbId being the same value as the year (because that's how Breedbase returns it from the GET endpoint).